refactored BasicCredentials
