+++ src/WebApiContrib/WebApiContrib.csproj
@@ -64,6 +64,7 @@
     <Compile Include="Caching\EntityTagKey.cs" />
     <Compile Include="Caching\TimedEntityTagHeaderValue.cs" />
     <Compile Include="Filters\BasicAuthenticationAttribute.cs" />
+    <Compile Include="Internal\BasicCredentials.cs" />
     <Compile Include="Internal\Extensions\ActionExtensions.cs" />
     <Compile Include="Internal\Extensions\IsInExtensions.cs" />
     <Compile Include="Internal\HttpHeaderNames.cs" />
